refactor: And separate gltf export works as well
