+++ valen-core/src/main/java/be/twofold/valen/core/util/Filenames.java
@@ -9,6 +9,10 @@ public final class Filenames {
         return index < 0 ? path : path.substring(index + 1);
     }
 
+    public static String fileNameWithoutExtension(String path) {
+        return removeExtension(fileName(path));
+    }
+
     public static String pathName(String path) {
         var index = path.lastIndexOf('/');
         return index < 0 ? "" : path.substring(0, index);
