+++ valen-export-gltf/src/main/java/be/twofold/valen/export/gltf/GlbModelExporter.java
@@ -2,6 +2,7 @@ package be.twofold.valen.export.gltf;
 
 import be.twofold.valen.core.export.*;
 import be.twofold.valen.core.geometry.*;
+import be.twofold.valen.core.util.*;
 import be.twofold.valen.export.gltf.mappers.*;
 import be.twofold.valen.gltf.*;
 
@@ -28,10 +29,18 @@ public final class GlbModelExporter implements Exporter<Model> {
 
     @Override
     public void export(Model model, Path path) throws IOException {
-        var writer = GltfWriter.createGlbWriter();
-        var modelMapper = new GltfModelMultiMapper(writer);
-        var rootNodeID = modelMapper.map(model);
-        writer.addScene(List.of(rootNodeID));
-        writer.write(path);
+        var outputDirectory = path.getParent();
+        var fileName = path.getFileName().toString();
+        var fileNameWithoutExtension = Filenames.removeExtension(fileName);
+
+        var binPath = outputDirectory.resolve(fileNameWithoutExtension + ".bin");
+        var imagePath = outputDirectory.resolve("_images");
+
+        try (var writer = GltfWriter.createSplitWriter(path, binPath, imagePath)) {
+//        try (var writer = GltfWriter.createGlbWriter(path)) {
+            var modelMapper = new GltfModelMultiMapper(writer);
+            var rootNodeID = modelMapper.map(model);
+            writer.addScene(List.of(rootNodeID));
+        }
     }
 }
