+++ valen-export-gltf/src/main/java/be/twofold/valen/export/gltf/GlbSceneExporter.java
@@ -2,7 +2,6 @@ package be.twofold.valen.export.gltf;
 
 import be.twofold.valen.core.export.*;
 import be.twofold.valen.core.scene.*;
-import be.twofold.valen.export.gltf.mappers.GltfUtils;
 import be.twofold.valen.export.gltf.mappers.*;
 import be.twofold.valen.gltf.*;
 import be.twofold.valen.gltf.model.mesh.*;
@@ -13,8 +12,6 @@ import java.nio.file.*;
 import java.util.*;
 
 public final class GlbSceneExporter implements Exporter<Scene> {
-    private final GltfWriter writer = GltfWriter.createGlbWriter();
-    private final GltfModelSingleMapper modelMapper = new GltfModelSingleMapper(writer, null);
 
     @Override
     public String getExtension() {
@@ -28,14 +25,16 @@ public final class GlbSceneExporter implements Exporter<Scene> {
 
     @Override
     public void export(Scene scene, Path path) throws IOException {
-        var instanceNodes = new ArrayList<NodeID>();
-        for (var instance : scene.instances()) {
-            modelMapper.map(instance.modelReference())
-                .ifPresent(meshID -> instanceNodes.add(mapInstance(instance, meshID)));
-        }
+        try (var writer = GltfWriter.createGlbWriter(path)) {
+            var modelMapper = new GltfModelSingleMapper(writer, null);
+            var instanceNodes = new ArrayList<NodeID>();
+            for (var instance : scene.instances()) {
+                modelMapper.map(instance.modelReference())
+                    .ifPresent(meshID -> instanceNodes.add(mapInstance(writer, instance, meshID)));
+            }
 
-        writer.addScene(instanceNodes);
-        writer.write(path);
+            writer.addScene(instanceNodes);
+        }
     }
 
     @Override
@@ -43,7 +42,7 @@ public final class GlbSceneExporter implements Exporter<Scene> {
         throw new UnsupportedOperationException("Cannot export a scene to a stream for now");
     }
 
-    private NodeID mapInstance(Instance instance, MeshID meshID) {
+    private NodeID mapInstance(GltfWriter writer, Instance instance, MeshID meshID) {
         return writer.addNode(ImmutableNode.builder()
             .name(Optional.ofNullable(instance.name()))
             .translation(GltfUtils.mapVector3(instance.translation()))
