--- valen-gltf/src/main/java/be/twofold/valen/gltf/GltfMode.java
@@ -1,6 +1,0 @@
-package be.twofold.valen.gltf;
-
-public enum GltfMode {
-    GLB,
-    GLTF_PLUS_BIN_PLUS_IMAGES,
-}
