--- valen-gltf/src/main/java/be/twofold/valen/gltf/GltfUtils.java
@@ -1,19 +1,0 @@
-package be.twofold.valen.gltf;
-
-import java.io.*;
-
-final class GltfUtils {
-    private GltfUtils() {
-    }
-
-    static int alignedLength(int length) {
-        return (length + 3) & ~3;
-    }
-
-    static void align(OutputStream output, int length, byte pad) throws IOException {
-        int padLength = alignedLength(length) - length;
-        for (int i = 0; i < padLength; i++) {
-            output.write(pad);
-        }
-    }
-}
